+++ test/requestContextPlugin.e2e.spec.js
@@ -1,65 +1,9 @@
-const fastify = require('fastify')
 const request = require('superagent')
-const { fastifyRequestContextPlugin } = require('../lib/requestContextPlugin')
+const {
+  initAppPostWithPrevalidation,
+  initAppPostWithAllPlugins,
+} = require('./internal/appInitializer')
 const { TestService } = require('./internal/testService')
-
-async function initAppPostWithPrevalidation(endpoint) {
-  const app = fastify({ logger: true })
-  app.register(fastifyRequestContextPlugin, { hook: 'preValidation' })
-
-  const preValidationFn = (req, reply, done) => {
-    const requestId = Number.parseInt(req.body.requestId)
-    req.requestContext.set('testKey', `testValue${requestId}`)
-    done()
-  }
-
-  app.route({
-    url: '/',
-    method: ['GET', 'POST'],
-    preValidation: preValidationFn,
-    handler: endpoint,
-  })
-
-  await app.listen(0)
-  return app
-}
-
-async function initAppPostWithAllPlugins(endpoint, requestHook) {
-  const app = fastify({ logger: true })
-  app.register(fastifyRequestContextPlugin, { hook: requestHook })
-
-  app.addHook('onRequest', (req, reply, done) => {
-    req.requestContext.set('onRequest', 'dummy')
-    done()
-  })
-
-  app.addHook('preParsing', (req, reply, payload, done) => {
-    req.requestContext.set('preParsing', 'dummy')
-    done(null, payload)
-  })
-
-  app.addHook('preValidation', (req, reply, done) => {
-    const requestId = Number.parseInt(req.body.requestId)
-    req.requestContext.set('preValidation', requestId)
-    req.requestContext.set('testKey', `testValue${requestId}`)
-    done()
-  })
-
-  app.addHook('preHandler', (req, reply, done) => {
-    const requestId = Number.parseInt(req.body.requestId)
-    req.requestContext.set('preHandler', requestId)
-    done()
-  })
-
-  app.route({
-    url: '/',
-    method: ['GET', 'POST'],
-    handler: endpoint,
-  })
-
-  await app.listen(0)
-  return app
-}
 
 describe('requestContextPlugin E2E', () => {
   let app
@@ -99,8 +43,8 @@ describe('requestContextPlugin E2E', () => {
             throw new Error(`Unexpected requestId: ${requestId}`)
           }
 
-          initAppPostWithPrevalidation(route).then((_app) => {
-            app = _app
+          app = initAppPostWithPrevalidation(route)
+          app.listen(0).then(() => {
             testService = new TestService(app)
             const { address, port } = app.server.address()
             const url = `${address}:${port}`
@@ -177,8 +121,9 @@ describe('requestContextPlugin E2E', () => {
             throw new Error(`Unexpected requestId: ${requestId}`)
           }
 
-          initAppPostWithAllPlugins(route, 'preValidation').then((_app) => {
-            app = _app
+          app = initAppPostWithAllPlugins(route, 'preValidation')
+
+          app.listen(0).then(() => {
             testService = new TestService(app)
             const { address, port } = app.server.address()
             const url = `${address}:${port}`
