+++ test/requestContextPlugin.spec.js
@@ -1,44 +1,9 @@
-const fastify = require('fastify')
-const { fastifyRequestContextPlugin } = require('../lib/requestContextPlugin')
+const {
+  initAppPost,
+  initAppPostWithPrevalidation,
+  initAppGet,
+} = require('./internal/appInitializer')
 const { TestService } = require('./internal/testService')
-
-function initAppGet(endpoint) {
-  const app = fastify({ logger: true })
-  app.register(fastifyRequestContextPlugin)
-
-  app.get('/', endpoint)
-
-  return app.ready()
-}
-
-function initAppPost(endpoint) {
-  const app = fastify({ logger: true })
-  app.register(fastifyRequestContextPlugin)
-
-  app.post('/', endpoint)
-
-  return app.ready()
-}
-
-function initAppPostWithPrevalidation(endpoint) {
-  const app = fastify({ logger: true })
-  app.register(fastifyRequestContextPlugin, { hook: 'preValidation' })
-
-  const preValidationFn = (req, reply, done) => {
-    const requestId = Number.parseInt(req.body.requestId)
-    req.requestContext.set('testKey', `testValue${requestId}`)
-    done()
-  }
-
-  app.route({
-    url: '/',
-    method: ['GET', 'POST'],
-    preValidation: preValidationFn,
-    handler: endpoint,
-  })
-
-  return app.ready()
-}
 
 describe('requestContextPlugin', () => {
   let app
@@ -79,37 +44,39 @@ describe('requestContextPlugin', () => {
             }
           }
 
-          initAppGet(route).then((_app) => {
-            app = _app
-            testService = new TestService(app)
-            const response1Promise = app
-              .inject()
-              .get('/')
-              .query({ requestId: 1 })
-              .end()
-              .then((response) => {
-                expect(response.json().storedValue).toBe('testValue1')
-                responseCounter++
-                if (responseCounter === 2) {
-                  resolveResponsePromise()
-                }
-              })
+          initAppGet(route)
+            .ready()
+            .then((_app) => {
+              app = _app
+              testService = new TestService(app)
+              const response1Promise = app
+                .inject()
+                .get('/')
+                .query({ requestId: 1 })
+                .end()
+                .then((response) => {
+                  expect(response.json().storedValue).toBe('testValue1')
+                  responseCounter++
+                  if (responseCounter === 2) {
+                    resolveResponsePromise()
+                  }
+                })
 
-            const response2Promise = app
-              .inject()
-              .get('/')
-              .query({ requestId: 2 })
-              .end()
-              .then((response) => {
-                expect(response.json().storedValue).toBe('testValue2')
-                responseCounter++
-                if (responseCounter === 2) {
-                  resolveResponsePromise()
-                }
-              })
+              const response2Promise = app
+                .inject()
+                .get('/')
+                .query({ requestId: 2 })
+                .end()
+                .then((response) => {
+                  expect(response.json().storedValue).toBe('testValue2')
+                  responseCounter++
+                  if (responseCounter === 2) {
+                    resolveResponsePromise()
+                  }
+                })
 
-            return Promise.all([response1Promise, response2Promise])
-          })
+              return Promise.all([response1Promise, response2Promise])
+            })
         })
 
         return promiseRequest1
@@ -152,37 +119,39 @@ describe('requestContextPlugin', () => {
             }
           }
 
-          initAppPost(route).then((_app) => {
-            app = _app
-            testService = new TestService(app)
-            const response1Promise = app
-              .inject()
-              .post('/')
-              .body({ requestId: 1 })
-              .end()
-              .then((response) => {
-                expect(response.json().storedValue).toBe('testValue1')
-                responseCounter++
-                if (responseCounter === 2) {
-                  resolveResponsePromise()
-                }
-              })
+          initAppPost(route)
+            .ready()
+            .then((_app) => {
+              app = _app
+              testService = new TestService(app)
+              const response1Promise = app
+                .inject()
+                .post('/')
+                .body({ requestId: 1 })
+                .end()
+                .then((response) => {
+                  expect(response.json().storedValue).toBe('testValue1')
+                  responseCounter++
+                  if (responseCounter === 2) {
+                    resolveResponsePromise()
+                  }
+                })
 
-            const response2Promise = app
-              .inject()
-              .post('/')
-              .body({ requestId: 2 })
-              .end()
-              .then((response) => {
-                expect(response.json().storedValue).toBe('testValue2')
-                responseCounter++
-                if (responseCounter === 2) {
-                  resolveResponsePromise()
-                }
-              })
+              const response2Promise = app
+                .inject()
+                .post('/')
+                .body({ requestId: 2 })
+                .end()
+                .then((response) => {
+                  expect(response.json().storedValue).toBe('testValue2')
+                  responseCounter++
+                  if (responseCounter === 2) {
+                    resolveResponsePromise()
+                  }
+                })
 
-            return Promise.all([response1Promise, response2Promise])
-          })
+              return Promise.all([response1Promise, response2Promise])
+            })
         })
 
         return promiseRequest1
@@ -224,37 +193,39 @@ describe('requestContextPlugin', () => {
             }
           }
 
-          initAppPostWithPrevalidation(route).then((_app) => {
-            app = _app
-            testService = new TestService(app)
-            const response1Promise = app
-              .inject()
-              .post('/')
-              .body({ requestId: 1 })
-              .end()
-              .then((response) => {
-                expect(response.json().storedValue).toBe('testValue1')
-                responseCounter++
-                if (responseCounter === 2) {
-                  resolveResponsePromise()
-                }
-              })
+          initAppPostWithPrevalidation(route)
+            .ready()
+            .then((_app) => {
+              app = _app
+              testService = new TestService(app)
+              const response1Promise = app
+                .inject()
+                .post('/')
+                .body({ requestId: 1 })
+                .end()
+                .then((response) => {
+                  expect(response.json().storedValue).toBe('testValue1')
+                  responseCounter++
+                  if (responseCounter === 2) {
+                    resolveResponsePromise()
+                  }
+                })
 
-            const response2Promise = app
-              .inject()
-              .post('/')
-              .body({ requestId: 2 })
-              .end()
-              .then((response) => {
-                expect(response.json().storedValue).toBe('testValue2')
-                responseCounter++
-                if (responseCounter === 2) {
-                  resolveResponsePromise()
-                }
-              })
+              const response2Promise = app
+                .inject()
+                .post('/')
+                .body({ requestId: 2 })
+                .end()
+                .then((response) => {
+                  expect(response.json().storedValue).toBe('testValue2')
+                  responseCounter++
+                  if (responseCounter === 2) {
+                    resolveResponsePromise()
+                  }
+                })
 
-            return Promise.all([response1Promise, response2Promise])
-          })
+              return Promise.all([response1Promise, response2Promise])
+            })
         })
 
         return promiseRequest1
